+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Make durability repair of lazurite tools with an ingot be based off of original ingredients rather than automatically 100%
 * Add power generation indicator to Stirling Engine
 * Add power draw indicator to Network Interface
+* Add sound to Brain crafting (washing?) recipe
+* Remove ability to place Magic Hat as a placeable block as it makes re-catching Presto really annoying
 
 ## 26.1.2.2